+++ src/main/java/com/perimeterx/internals/cookie/cookieparsers/HeaderParser.java
@@ -48,6 +48,9 @@ public abstract class HeaderParser {
     }
 
     public DataEnrichmentCookie getRawDataEnrichmentCookie(List<RawCookieData> rawCookies, String cookieKey) {
+        return getRawDataEnrichmentCookie(rawCookies, Collections.singletonList(cookieKey));
+    }
+    public DataEnrichmentCookie getRawDataEnrichmentCookie(List<RawCookieData> rawCookies, List<String> cookieKeys) {
         ObjectMapper mapper = new ObjectMapper();
         DataEnrichmentCookie dataEnrichmentCookie = new DataEnrichmentCookie(mapper.createObjectNode(), false);
         RawCookieData rawDataEnrichmentCookie = null;
@@ -67,7 +70,8 @@ public abstract class HeaderParser {
             String hmac = cookiePayloadArray[0];
             String encodedPayload = cookiePayloadArray[1];
 
-            boolean isValid = HMACUtils.isHMACValid(encodedPayload, hmac, cookieKey, logger);
+            boolean isValid = cookieKeys.stream()
+                    .anyMatch(cookieKey -> HMACUtils.isHMACValid(encodedPayload, hmac, cookieKey, logger));
             dataEnrichmentCookie.setValid(isValid);
 
             byte[] decodedPayload = Base64.decode(encodedPayload);